+++ src/main/java/yapp/bestFriend/service/v1/ProductService.java
@@ -1,4 +1,4 @@
-package yapp.bestFriend.service;
+package yapp.bestFriend.service.v1;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -120,5 +120,22 @@ public class ProductService {
             else return DefaultRes.response(HttpStatus.OK.value(), "삭제 실패(사용자 정보 없음)");
         }
         else return DefaultRes.response(HttpStatus.OK.value(), "삭제 실패(절약 정보 없음)");
+    }
+
+    /************************ version 1.1 *************************/
+
+    public DefaultRes<List<SimpleProductResponse>> getProductListV1pt1(Long userId, String recordYmd) {
+        return null;
+    }
+
+    public DefaultRes<List<SimpleProductResponse>> createProductV1pt1(Long userId, yapp.bestFriend.model.dto.request.v1pt1.product.CreateProductRequest request) {
+        Optional<User> user = userRepository.findById(userId);
+
+        if(user.isPresent()){
+            user.get().getProductList().add(productRepository.save(request.toEntity(user.get())));
+
+            return DefaultRes.response(HttpStatus.OK.value(), "등록 성공");
+        }
+        else return DefaultRes.response(HttpStatus.OK.value(), "등록 실패(사용자 정보 없음)");
     }
 }
