+++ src/main/java/yapp/bestFriend/service/v1/auth/KakaoOauth.java
@@ -1,4 +1,4 @@
-package yapp.bestFriend.service.auth;
+package yapp.bestFriend.service.v1.auth;
 
 import com.google.gson.JsonParser;
 import lombok.RequiredArgsConstructor;
