+++ src/test/java/yapp/bestFriend/service/alarm/AlramServiceTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.junit.jupiter.MockitoExtension;
+import yapp.bestFriend.service.v1.alarm.AlramService;
 
 import java.time.LocalDateTime;
