+++ src/test/java/yapp/bestFriend/service/savingRecord/SavingRecordServiceTest.java
@@ -17,6 +17,7 @@ import yapp.bestFriend.repository.ProductRepository;
 import yapp.bestFriend.repository.SavingRecordRepository;
 import yapp.bestFriend.repository.SavingRecordRepositoryCustom;
 import yapp.bestFriend.repository.UserRepository;
+import yapp.bestFriend.service.v1.savingRecord.SavingRecordService;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
